# Batch create Sections and Subsections in _build_nofo function